Fix bug in definition of psi
The angle psi is defined as psi=mu-thetap, where mu is the angle between
the guiding-center velocity vector V and the observer, and thetap is the
pitch angle of the particle (defined above w.r.t. to the guiding-center
velocity, and thus always >= 0). In SOFT, however, the pitch angle is
defined w.r.t. to the magnetic field vector, and so cos(thetap) may be
negative if the guiding-center is travelling in the anti-parallel
direction. Hence, we must correct for this when calculating cos(psi). We
can do this by multiplying cos(thetap) by V.B/(VB) (since sin(thetap) is
defined such that it is always positive).